Refactor creating basic GraphQL types for Scalar DB
