--- graphql/src/main/java/com/scalar/db/graphql/schema/DeleteConditionType.java
@@ -1,6 +1,0 @@
-package com.scalar.db.graphql.schema;
-
-public enum DeleteConditionType {
-  DeleteIf,
-  DeleteIfExists
-}
