--- graphql/src/main/java/com/scalar/db/graphql/schema/PutConditionType.java
@@ -1,7 +1,0 @@
-package com.scalar.db.graphql.schema;
-
-public enum PutConditionType {
-  PutIf,
-  PutIfExists,
-  PutIfNotExists
-}
